Refactor napps view into frame widget
