--- src/types/account.ts
@@ -1,7 +1,0 @@
-import { z } from 'zod'
-
-export const accountDataSchema = z.object({
-  name: z.string()
-})
-
-export type AccountData = z.infer<typeof accountDataSchema>
